+++ ArraySim/Isotrope.h
@@ -4,12 +4,12 @@
 class Isotrope : public Antenna
 {
 private:
-	float taper = 1;
-	float phase = 0;
+	double taper = 1;
+	double phase = 0;
 
 public:
-	float get_gain(float el, float az, float frequency);//in radians, samples from linear, not logarithmic, gain pattern
-	void set_taper(float taper);//taper is a static multiplier to use for the gain
-	void set_phase(float phase);//the phase offset for array steering
-	float get_phase(float el, float az, float frequency);//basic model would ignore arguments
+	double get_gain(double el, double az, double frequency);//in radians, samples from linear, not logarithmic, gain pattern
+	void set_taper(double taper);//taper is a static multiplier to use for the gain
+	void set_phase(double phase);//the phase offset for array steering
+	double get_phase(double el, double az, double frequency);//basic model would ignore arguments
 };
